requirements.txt
Django==5.2.2
torch
torchvision
pillow